+++ test/python/transpiler/test_pass_scheduler.py
@@ -22,10 +22,9 @@ import unittest.mock
 import sys
 
 from qiskit import QuantumRegister, QuantumCircuit
-from qiskit.transpiler import PassManager
+from qiskit.transpiler import PassManager, TranspilerAccessError, TranspilerError
 from qiskit.compiler import transpile
-from qiskit.transpiler import TranspilerAccessError, TranspilerError
-from qiskit.transpiler.passmanager import DoWhileController, ConditionalController, \
+from qiskit.transpiler.runningpassmanager import DoWhileController, ConditionalController, \
     FlowController, FlowControllerLinear
 from qiskit.test import QiskitTestCase
 from ._dummy_passes import (PassA_TP_NR_NP, PassB_TP_RA_PA, PassC_TP_RA_PA,
